moving some functionality to file util
